Change from the tabletop demo to real world control system
Increased the trigger and give up times to minutes from seconds.
Switch on real pumps and valves connected via solid state relais
to GPIO pins.

Changed the control logic a bit: keep the pump running if the flow is
above 45 C to give it some time to heat up to loading temperature when
the radiation is weak